Hotfix 5.0.1: Int to Long migration, Database Safety, Core Bugs

+++ core/common/src/commonMain/kotlin/com/kazemieh/common/model/Transaction.kt
@@ -7,8 +7,8 @@ import kotlin.time.Clock
 @Serializable
 data class Transaction(
     val id: Long,
-    val amount: Int,
-    val amountTransfer: Int = 0,
+    val amount: Long,
+    val amountTransfer: Long = 0L,
     val categoryId: Long,
     val sourceId: Long,
     val sourceEndId: Long? = null,
